(Minor) Monk Rework (#2208)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Frees Monk patron slots from Ravox to all temple patrons - and gives
them requisite amulet. Also gives monk access to churchling-level
miracles, prayers and devotion systems, unique to their patron.

Adjusts speed by +1, strength by -1, ~with a few species-specific
balance changes for the role~.


## Why It's Good For The Game

The goal of this PR is to open Monk to more roleplay possibilities, and
to bring the role closer to DND equivalents with less focuses on
grappling and STR gimmicks (barbarian exists for this purpose), and more
faith systems.

A few images;

![image](https://github.com/user-attachments/assets/9cec5b63-329a-4a30-abef-3cf3f0a22837)

![image](https://github.com/user-attachments/assets/90f4fc43-02f5-43e2-9118-d875670bdeb4)

![monk1](https://github.com/user-attachments/assets/35fb2cd8-d0cf-421f-b319-5fa2e35a89b0)

![image](https://github.com/user-attachments/assets/437039e9-f054-4293-8407-331359544f94)


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>