Hollow Verderer: A new, Hollowkin and Humen Mercenary Subclass! (#2295)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Adds a Bardiche-swinging, rust-armor-wearing, more survivalist fighter
to the Mercenary subclasses for Hollowkin (and Humen). This class is
focused on the identity of the people of Amber Hollow as very scrappy
and more outdoorsy than most of the other civilizations, being forced
into using worse gear than most but pulling together everything they can
to survive. As a class, this should play fairly decently, but using
skills to improve leather gear or doing jobs in order to fund non-rusted
armor is going to be a big incentive.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Hollowkin currently lack ANY special subclasses like other races get,
and with Mercenary this is especially pertinent- You can ONLY become a
Boltslinger as a Hollowkin! This class, and hopefully other classes I'll
begin work on soon, will serve to pad out the options for Hollowkin
players, allowing more diversity in play than just "shoot crossbow make
mammons" for mercenary Hollowkin. Humen are also granted the class
option because there's lore that the people of Amber Hollow work with
bandits which could very easily be Humen Matthiosans.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->