Personal Objectives DLC 2: Adopt orphans, scorn inhumen and make an ultimate sacrifice (#2317)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR adds a new personal objective to each god, along with general
improvements and tweaks for the old ones and personal objectives in
general.

**General:**

- Better folder organization of the relevant events and objectives, so
they are sorted by each god
- Lots of minor tweaks, renamings and code improvements for the older
personal objectives, including make use of triumph_count variable
- Activated wise trees will now try to protect normal trees around them,
not just themselves
- Small point gain ponus for personal event track, as we have a lot more
of them now

**New Objectives:**
- Astrata - Spit on 2 inhumen
- Noc - Take an apprentice
- Eora - Adopt an orphan
- Malum - Spend enough energy working
- Dendor - Butcher 2 animals
- Ravox - Make an ultimate sacrifice to give another person a second
chance to live
- Necra - Hear the cries of the dead in the church while wearing her
amulet
- Abyssor - If the Abyssor is calm, take a bath while wearing his
amulet. If the Abyssor is mad, splash water on someone
- Pestra - Take enough pain from others upon yourself
- Xylix - Win a game against Xylix, where rules change on the fly and
cheating is encouraged
- Zizo - Kick a man in the groin
- Graggar - Perform a blood rite
- Baotha - Find 10 flaws of others
- Matthios - Rob 2 graves

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Better organization and variety for personal objectives.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>