Actually modernize CI workflow (#2275)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
There were pulls made before us or after forking that added the backend
for modernising workflows and such but didn't actually update the CI
suite or add workflows so, lets do that.

We failing CI with this one :fire::fire::fire:

Basically will port every pr from 2021 to 2025 that wasn't already that
changes ci_suite.yml that is useful

Want https://github.com/tgstation/tgstation/pull/89180 but can't do :(

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More CI checks for better prs, less time spent running that CI.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->